Create a game function .
Create a game function that doesn't take any parameter and let the user and the computer to play a serie of 5 games
Call the Playround () function 5 times using a loop
check the return value of the PlayRound () to increment the score and print the winner.
Comare the scores at the end the print the winner.